Large refactor after moving files around and adding linter
