+++ .gitignore
@@ -58,7 +58,7 @@ typings/
 .yarn-integrity
 
 # dotenv environment variables file
-.env
+.env*
 dev.env
 
 #Exclude tsc generated files
