+++ tslint.json
@@ -1,0 +1,3 @@
+{
+  "extends": "gts/tslint.json"
+}
